+++ excelerator/__init__.py
@@ -2,7 +2,8 @@
 Library for reading tabular data from excel in a very forgiving manner.
 """
 
-__version__ = "0.5"
+__version__ = "0.6"
 
 
 from excelerator.main.table_reader import TableReader
+from excelerator.main.field import Field